User, Website and Password all return to main now!
Took me ages to wrap my head around pointer pass function nonsense, think I'm getting it thought. Not only wanted to learn how it worked, but also would like to be able to append the 3 strings to a txt file or something similar. So figured I'd have to be able to access the strings in main, not just print them within their own respective functions.